Move App file
